Move handler functions to a separated file
